% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomise.R
\name{biomise}
\alias{biomise}
\title{A wrapper for all the crest functions.}
\usage{
biomise(s, pol2pft, pft2biome)
}
\arguments{
\item{s}{.}

\item{pol2pft}{.}

\item{pft2biome}{.}
}
\value{
A .
}
\description{
Runs all the different steps of a CREST reconstruction in one function.
}
\examples{
1:5

}
