% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMD.R
\name{EMD}
\alias{EMD}
\title{A wrapper for all the crest functions.}
\usage{
EMD(
  x,
  y,
  weight.m = matrix(rep(1, length(x)^2), ncol = length(x), byrow = TRUE) - diag(1,
    length(x), length(x))
)
}
\arguments{
\item{site_info}{A vector containing the coordinates of the study site.}
}
\value{
A \code{\link{crestObj}} containing the reconstructions.
}
\description{
Runs all the different steps of a CREST reconstruction in one function.
}
\examples{

}
