% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft2biome.R
\name{pft2biome}
\alias{pft2biome}
\title{A wrapper for all the crest functions.}
\usage{
pft2biome(s, pft2biome)
}
\arguments{
\item{s}{.}

\item{pft2biome}{.}
}
\value{
A .
}
\description{
Runs all the different steps of a CREST reconstruction in one function.
}
\examples{
1:5

}
