% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMD.R
\name{EMD}
\alias{EMD}
\title{A wrapper for all the crest functions.}
\usage{
EMD(
  x,
  y,
  weight.m = matrix(rep(1, length(x)^2), ncol = length(x), byrow = TRUE) - diag(1,
    length(x), length(x))
)
}
\arguments{
\item{x, y}{The two vectors to compare}

\item{weight.m}{Matrix of weights. Values should be 'numeric'.}
}
\value{
The EMD between \code{x} and \code{y}.
}
\description{
Calculates the EMD for two vectors \code{x} and \code{y}.
}
\examples{
EMD(1:5, 6:10)
m <- matrix(1:25, ncol=5)
for(i in 1:5) m[i,i] <- 0
EMD(1:5, 6:10, weight.m=m)
\dontrun{
  EMD(1:5, 1:6)
}

}
