% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollen2pft.R
\name{pollen2pft}
\alias{pollen2pft}
\title{A wrapper for all the crest functions.}
\usage{
pollen2pft(s, pol2pft)
}
\arguments{
\item{site_info}{A vector containing the coordinates of the study site.}
}
\value{
A \code{\link{crestObj}} containing the reconstructions.
}
\description{
Runs all the different steps of a CREST reconstruction in one function.
}
\examples{

}
